rename: move auth.interface.ts to auth/services/auth
