+++ backend/src/auth/controllers/auth/auth.controller.ts
@@ -11,12 +11,12 @@ import {
 import { FtAuthGurad } from 'src/auth/guards/ft-auth.guard';
 import { JwtAccessAuthGuard } from 'src/auth/guards/jwt.access-auth.guard';
 import { JwtRefreshAuthGuard } from 'src/auth/guards/jwt.refresh-auth.guard';
-import { AuthService } from 'src/auth/services/auth/auth.service';
+import { IAuthService } from 'src/auth/services/auth/auth.interface';
 import { Cookies } from 'src/utils/types';
 
 @Controller('auth')
 export class AuthController {
-  constructor(@Inject('AUTH_SERVICE') private authService: AuthService) {}
+  constructor(@Inject('AUTH_SERVICE') private authService: IAuthService) {}
 
   @UseGuards(JwtRefreshAuthGuard)
   @Redirect('http://localhost:8000/Home', 301)
